example--prompt/pyproject.toml
[tool.poetry]
name = "example-prompt"
version = "0.0.1"
description = ""
authors = ["Andre Pasold <andre.pasold@totvs.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^3.0.3"
python-dotenv = "^1.0.1"
langfuse = "^2.38.1"
openai = "^1.35.15"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
